backend-python/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
motor>=3.3.0
pymongo>=4.6.0

# Image processing
pillow>=10.2.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Google Cloud
google-cloud-vision>=3.4.0
google-cloud-documentai>=2.20.0
google-auth>=2.23.0
google-generativeai>=0.3.0

# AWS
boto3>=1.29.0

# HTTP
requests>=2.31.0
httpx>=0.25.0
